fix: 修复编译错误 - 使用nativeCanvas绘制文字、添加material-icons-extended依赖、修复返回类型

+++ app/src/main/java/com/satfinder/app/satellite/SatelliteCalculator.kt
@@ -136,8 +136,8 @@ object SatelliteCalculator {
         if (dEl < -verticalFOV / 2 || dEl > verticalFOV / 2) return null
 
         // 映射到屏幕坐标
-        val x = screenWidth / 2f + (dAz / (horizontalFOV / 2)) * (screenWidth / 2f)
-        val y = screenHeight / 2f - (dEl / (verticalFOV / 2)) * (screenHeight / 2f)
+        val x = (screenWidth / 2.0 + (dAz / (horizontalFOV / 2.0)) * (screenWidth / 2.0)).toFloat()
+        val y = (screenHeight / 2.0 - (dEl / (verticalFOV / 2.0)) * (screenHeight / 2.0)).toFloat()
 
         return Pair(x, y)
     }
